+++ propagation/baggage_test.go
@@ -17,7 +17,6 @@ package propagation_test
 import (
 	"context"
 	"net/http"
-	"net/url"
 	"strings"
 	"testing"
 
@@ -47,7 +46,7 @@ func (m member) Member(t *testing.T) baggage.Member {
 		}
 		props = append(props, p)
 	}
-	bMember, err := baggage.NewMember(m.Key, url.PathEscape(m.Value), props...)
+	bMember, err := baggage.NewMember(m.Key, m.Value, props...)
 	if err != nil {
 		t.Fatal(err)
 	}